Переименован адаптер дз Сделаны обработчики нажатия для флагов
